Create depth-sensing-JSN-SRO4_Gen3_RK-SD-GPS-temp-integration-attempt1
I have copied my code into this file! I attempted to integrate the JSN-SRO4_Gen3_RK code with the preexisting code, but I'm not sure that it will fully work. My computer is having some issues with Visual Studio Code, so I'm not able to flash the code there. I tried the Particle Workbench Web IDE, but you have to have a device to compile/flash the code, so I haven't been able to compile the code and check for any specific errors yet. Please let me know if there are any errors or issues that arise with the code! Thank you!!